+++ src/main/java/sonar/fluxnetworks/common/item/ItemFluxConnector.java
@@ -1,10 +1,5 @@
 package sonar.fluxnetworks.common.item;
 
-import sonar.fluxnetworks.api.translate.FluxTranslate;
-import sonar.fluxnetworks.api.utils.EnergyType;
-import sonar.fluxnetworks.client.FluxColorHandler;
-import sonar.fluxnetworks.common.data.FluxNetworkData;
-import sonar.fluxnetworks.common.core.FluxUtils;
 import net.minecraft.block.Block;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.item.ItemBlock;
@@ -14,6 +9,11 @@ import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+import sonar.fluxnetworks.api.translate.FluxTranslate;
+import sonar.fluxnetworks.api.utils.EnergyType;
+import sonar.fluxnetworks.client.FluxColorHandler;
+import sonar.fluxnetworks.common.core.FluxUtils;
+import sonar.fluxnetworks.common.data.FluxNetworkData;
 
 import javax.annotation.Nullable;
 import java.text.NumberFormat;
@@ -34,7 +34,7 @@ public class ItemFluxConnector extends ItemBlock {
     @Override
     public String getItemStackDisplayName(ItemStack stack) {
         NBTTagCompound tag = stack.getSubCompound(FluxUtils.FLUX_DATA);
-        if(tag != null && tag.hasKey(CUSTOM_NAME)) {
+        if (tag != null && tag.hasKey(CUSTOM_NAME)) {
             return tag.getString(CUSTOM_NAME);
         }
         return super.getItemStackDisplayName(stack);
@@ -44,12 +44,12 @@ public class ItemFluxConnector extends ItemBlock {
     @SideOnly(Side.CLIENT)
     public void addInformation(ItemStack stack, @Nullable World worldIn, List<String> tooltip, ITooltipFlag flagIn) {
         NBTTagCompound tag = stack.getSubCompound(FluxUtils.FLUX_DATA);
-        if(tag != null) {
+        if (tag != null) {
             tooltip.add(FluxTranslate.NETWORK_FULL_NAME.t() + ": " + TextFormatting.WHITE + FluxColorHandler.getOrRequestNetworkName(tag.getInteger(FluxNetworkData.NETWORK_ID)));
             tooltip.add(FluxTranslate.TRANSFER_LIMIT.t() + ": " + TextFormatting.WHITE + FluxUtils.format(tag.getLong(LIMIT), FluxUtils.TypeNumberFormat.COMMAS, EnergyType.RF.getStorageSuffix()));
             tooltip.add(FluxTranslate.PRIORITY.t() + ": " + TextFormatting.WHITE + tag.getInteger(PRIORITY));
-            if(tag.hasKey("energy")) {
-                tooltip.add(FluxTranslate.ENERGY_STORED.t() + ": " + TextFormatting.WHITE + NumberFormat.getInstance().format(tag.getInteger("energy")) + "RF");
+            if (tag.hasKey("energy")) {
+                tooltip.add(FluxTranslate.ENERGY_STORED.t() + ": " + TextFormatting.WHITE + NumberFormat.getInstance().format(tag.getLong("energy")) + "RF");
             } else {
                 tooltip.add(FluxTranslate.INTERNAL_BUFFER.t() + ": " + TextFormatting.WHITE + FluxUtils.format(tag.getLong("buffer"), FluxUtils.TypeNumberFormat.COMMAS, "RF"));
             }
